% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Ask ChatGPT Completions Endpoint}
\usage{
ask_chatgpt(prompt, openai_api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{prompt}{The prompt to generate completions for.}

\item{openai_api_key}{OpenAI's API key.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
Ask ChatGPT Completions Endpoint
}
\examples{
\dontrun{
ask_chatgpt("What do you think about R language?")
}
}
