% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prompt_content.R
\name{build_prompt_content}
\alias{build_prompt_content}
\title{Build Prompt Content}
\usage{
build_prompt_content(question, images)
}
\arguments{
\item{question}{The question to ask ChatGPT.}

\item{images}{A list of images to attach to the question. It could be a list of URLs or paths.}
}
\description{
Build Prompt Content
}
