% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt-package.R
\docType{package}
\name{chatgpt-package}
\alias{chatgpt}
\alias{chatgpt-package}
\title{'OpenAI's 'ChatGPT' <https://chat.openai.com/> coding assistant for 'RStudio'. A set
of functions and 'RStudio' addins that aim to help the R developer in tedious coding tasks.}
\description{
'OpenAI's 'ChatGPT' \url{https://chat.openai.com/} coding assistant for 'RStudio'. A set of functions and 'RStudio' addins that aim to help the R developer in tedious coding tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcrodriguez1989/chatgpt}
  \item Report bugs at \url{https://github.com/jcrodriguez1989/chatgpt/issues}
}

}
\author{
\strong{Maintainer}: Juan Cruz Rodriguez \email{jcrodriguez@unc.edu.ar}

}
