% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{chats}
\alias{chats}
\title{Send a request to the chat endpoint}
\usage{
chats(
  messages,
  payload = list(model = "gpt-3.5-turbo"),
  token = config::get()$openai
)
}
\arguments{
\item{payload}{the POST request payload given}

\item{token}{api token}

\item{prompt}{a text prompt}
}
\description{
Send a request to the chat endpoint
}
