% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_code.R
\name{comment_code}
\alias{comment_code}
\title{ChatGPT: Comment Code}
\usage{
comment_code(code)
}
\arguments{
\item{code}{The code to be commented by ChatGPT.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Comment Code
}
\examples{
\dontrun{
cat(comment_code("for (i in 1:10) {\n  print(i ** 2)\n}"))
}

}
