% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_code.R
\name{complete_code}
\alias{complete_code}
\title{ChatGPT: Complete Code}
\usage{
complete_code(code)
}
\arguments{
\item{code}{The code to be completed by ChatGPT.}
}
\value{
The completed R code.
}
\description{
ChatGPT: Complete Code
}
