% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{completions}
\alias{completions}
\title{Send a request to the completions endpoint}
\usage{
completions(
  prompt,
  payload = list(model = "text-davinci-003"),
  token = config::get()$openai
)
}
\arguments{
\item{prompt}{a text prompt}

\item{payload}{the POST request payload given}

\item{token}{api token}
}
\description{
Send a request to the completions endpoint
}
