% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unit_tests.R
\name{create_unit_tests}
\alias{create_unit_tests}
\title{ChatGPT: Create Unit Tests}
\usage{
create_unit_tests(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code for which to create unit tests by ChatGPT. If not provided, it will use
what's copied on the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
Create `{testthat}` test cases for the code.
}
\examples{
\dontrun{
cat(create_unit_tests("squared_numbers <- function(numbers) {\n  numbers ^ 2\n}"))
}

}
