% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variable_name.R
\name{create_variable_name}
\alias{create_variable_name}
\title{ChatGPT: Create Variable Name}
\usage{
create_variable_name(code)
}
\arguments{
\item{code}{The code for which to give a variable name to its result.}
}
\description{
ChatGPT: Create Variable Name
}
