% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_prompt.R
\name{custom_prompt}
\alias{custom_prompt}
\title{ChatGPT: Custom Prompt}
\usage{
custom_prompt(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code to be appended to your custom prompt prior to sending to
ChatGPT. If not provided, it will use what's copied on the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Custom Prompt
}
