% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_code.R
\name{explain_code}
\alias{explain_code}
\title{ChatGPT: Explain Code}
\usage{
explain_code(code)
}
\arguments{
\item{code}{The code to be explained by ChatGPT.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Explain Code
}
\examples{
\dontrun{
cat(explain_code("for (i in 1:10) {\n  print(i ** 2)\n}"))
}

}
