% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_issues_in_code.R
\name{find_issues_in_code}
\alias{find_issues_in_code}
\title{ChatGPT: Find Issues in Code}
\usage{
find_issues_in_code(code)
}
\arguments{
\item{code}{The code to be analyzed by ChatGPT.}
}
\description{
ChatGPT: Find Issues in Code
}
