% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_image.R
\name{generate_image}
\alias{generate_image}
\title{Generate an Image With DALL-E 3}
\usage{
generate_image(
  prompt,
  out_file = tempfile(fileext = ".png"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{prompt}{The prompt for image generation.}

\item{out_file}{The path where to save the generated image.}

\item{openai_api_key}{OpenAI's API key.}
}
\description{
Generate an Image With DALL-E 3
}
