% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chat_session.R
\name{get_chat_session}
\alias{get_chat_session}
\title{Get Chat Session}
\usage{
get_chat_session(session_id = "1")
}
\arguments{
\item{session_id}{The ID of the session to be used. If `NULL`, it will return an empty session.}
}
\description{
Get Chat Session
}
