% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_entity.R
\name{gpt_entity}
\alias{gpt_entity}
\title{Get GPT Entity Recognition Result}
\usage{
result = gpt_entity(text)
}
\arguments{
\item{text}{a message}
}
\description{
Get an entitiy recognition result from ChatGPT
}
