% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_get_completions.R
\name{gpt_get_completions}
\alias{gpt_get_completions}
\title{Get GPT Completions Endpoint}
\usage{
gpt_get_completions(prompt, openai_api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{prompt}{The prompt to generate completions for.}

\item{openai_api_key}{OpenAI's API key.}
}
\description{
Get GPT Completions Endpoint
}
