% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{gpt_request}
\alias{gpt_request}
\title{Send a request to openai}
\usage{
gpt_request(base_url, payload, token = config::get()$openai)
}
\arguments{
\item{base_url}{The url fo the endpoint}

\item{payload}{the payload, an R list to be converted to JSON}

\item{token}{your API auth token}
}
\description{
Send a request to openai
}
