% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests_stream.R
\name{gpt_request_stream}
\alias{gpt_request_stream}
\title{Perform a streaming GPT API request}
\usage{
gpt_request_stream(base_url, callback, payload, token = config::get()$openai)
}
\arguments{
\item{base_url}{The endpoint URL for the GPT API request.}

\item{callback}{A callback function that defines how to process the data
as it streams in. This function should take a single argument, which
will be the raw vector of bytes received from the stream.}

\item{payload}{A list containing the data to be sent as the body of the POST
request. This will be converted to JSON format.}

\item{token}{Your OpenAI API token used for authentication.
Default is to retrieve from your config settings.}
}
\value{
An object of class \code{c("curl_response", "list")} that includes
information about the request and the status of the streaming data.
}
\description{
This function sends a streaming request to the GPT API and handles the response
using a callback function. It is designed to keep the connection open and
process the data as it streams in.
}
