% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_sent.R
\name{gpt_sent}
\alias{gpt_sent}
\title{Get GPT Sentiment Analysis Result}
\usage{
result = gpt_sent(text) %>% as.numeric()
}
\arguments{
\item{text}{a message}
}
\description{
Get a sentiment analysis reuslt from ChatGPT
}
