% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{ChatGPT: List Models}
\usage{
list_models(openai_api_key = Sys.getenv("OPENAI_API_KEY"))
}
\arguments{
\item{openai_api_key}{OpenAI's API key.}
}
\value{
A data.frame with the available models to be used by OpenAI's API.
}
\description{
ChatGPT: List Models
}
\examples{
\dontrun{
list_models()
}

}
