% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_code.R
\name{optimize_code}
\alias{optimize_code}
\title{ChatGPT: Optimize Code}
\usage{
optimize_code(code)
}
\arguments{
\item{code}{The code to be optimized by ChatGPT.}
}
\description{
ChatGPT: Optimize Code
}
