% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_code.R
\name{refactor_code}
\alias{refactor_code}
\title{ChatGPT: Refactor Code}
\usage{
refactor_code(code)
}
\arguments{
\item{code}{The code to be refactored by ChatGPT.}
}
\description{
ChatGPT: Refactor Code
}
