% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{run_addin}
\alias{run_addin}
\title{Run a ChatGPT RStudio Addin}
\usage{
run_addin(addin_name)
}
\arguments{
\item{addin_name}{The name of the adding to execute.}
}
\description{
Run a ChatGPT RStudio Addin
}
