% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{start_chat}
\alias{start_chat}
\title{Start a chatgpt session in the R terminal}
\usage{
start_chat(prompt)
}
\arguments{
\item{prompt}{the beginning prompt for your session}
}
\description{
Start a chatgpt session in the R terminal
}
