% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Ask ChatGPT}
\usage{
ask_chatgpt(
  question,
  session_id = "1",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{question}{The question to ask ChatGPT.}

\item{session_id}{The ID of the session to be used. We can have different conversations by using
different session IDs.}

\item{openai_api_key}{OpenAI's API key.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
Note: See also `reset_chat_session`.
}
\examples{
\dontrun{
cat(ask_chatgpt("What do you think about R language?"))
}

}
