% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chat_r6.R
\name{Chat}
\alias{Chat}
\title{Createa a running chat dialogue}
\description{
This R6 class stores chat messages to be used in the future  calls, so gpt can retain 'memory' of the conversation. Additionally there are some helpers to display token usage and to reset the chat.
}
\examples{
chat = Chat$new('List ingredients in spaghetti')

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Chat-reset}{\code{Chat$reset()}}
\item \href{#method-Chat-new}{\code{Chat$new()}}
\item \href{#method-Chat-chat}{\code{Chat$chat()}}
\item \href{#method-Chat-usage}{\code{Chat$usage()}}
\item \href{#method-Chat-clone}{\code{Chat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chat-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Chat-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chat$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chat-new"></a>}}
\if{latex}{\out{\hypertarget{method-Chat-new}{}}}
\subsection{Method \code{new()}}{
create a new chat session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chat$new(
  message,
  payload = list(model = "gpt-3.5-turbo"),
  token = config::get()$openai
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{the message you'd like to send}

\item{\code{payload}}{additional post body options, usually the model}

\item{\code{token}}{your openai auth token}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chat-chat"></a>}}
\if{latex}{\out{\hypertarget{method-Chat-chat}{}}}
\subsection{Method \code{chat()}}{
create a new chat session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chat$chat(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{the message you'd like to send}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chat-usage"></a>}}
\if{latex}{\out{\hypertarget{method-Chat-usage}{}}}
\subsection{Method \code{usage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chat$usage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Chat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Chat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Chat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
