% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_code.R
\name{complete_code}
\alias{complete_code}
\title{ChatGPT: Complete Code}
\usage{
complete_code(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code to be completed by ChatGPT. If not provided, it will use what's copied on
the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Complete Code
}
\examples{
\dontrun{
cat(complete_code("# A function to square each element of a vector\nsquare_each <- function("))
}

}
