% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variable_name.R
\name{create_variable_name}
\alias{create_variable_name}
\title{ChatGPT: Create Variable Name}
\usage{
create_variable_name(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code for which to give a variable name to its result. If not provided, it will
use what's copied on the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Create Variable Name
}
\examples{
\dontrun{
cat(create_variable_name("sapply(1:10, function(i) i ** 2)"))
}

}
