% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_issues_in_code.R
\name{find_issues_in_code}
\alias{find_issues_in_code}
\title{ChatGPT: Find Issues in Code}
\usage{
find_issues_in_code(code = clipr::read_clip(allow_non_interactive = TRUE))
}
\arguments{
\item{code}{The code to be analyzed by ChatGPT. If not provided, it will use what's copied on
the clipboard.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Find Issues in Code
}
\examples{
\dontrun{
cat(find_issues_in_code("i <- 0\nwhile (i < 0) {\n  i <- i - 1\n}"))
}

}
