% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_code.R
\name{optimize_code}
\alias{optimize_code}
\title{ChatGPT: Optimize Code}
\usage{
optimize_code(code)
}
\arguments{
\item{code}{The code to be optimized by ChatGPT.}
}
\value{
A character value with the response generated by ChatGPT.
}
\description{
ChatGPT: Optimize Code
}
\examples{
\dontrun{
cat(optimize_code("i <- 10\nwhile (i > 0) {\n  i <- i - 1\n  print(i)\n}"))
}

}
